+++ src/FleetManager.communication/Requests/RequestRentJson.cs
@@ -7,7 +7,7 @@
         public long ClientId { get; set; }
         public long VehicleId { get; set; }
         public long RentalPlanId {get ; set ;}
-        public decimal IncludedKm { get; set; }
+        public long IncludedKm { get; set; }
         public DateTime StartDate { get; set; }
         public DateTime? EndDate { get; set; }
     }
