+++ src/FleetManager.communication/Requests/RequestUpdateRentJson.cs
@@ -2,7 +2,7 @@
 {
     public class RequestUpdateRentJson
     {
-        public decimal IncludedKm { get; set; }
+        public long IncludedKm { get; set; }
         public DateTime StartDate { get; set; }
         public DateTime EndDate { get; set; }
     }
